requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
requests==2.31.0
python-dotenv==1.0.1
loguru==0.7.2
streamlit==1.35.0
matplotlib==3.9.2
pandas==2.2.2
numpy==2.1.2
pyarrow==17.0.0